Initial 2fa support
This is preliminary 2fa support. It depends on some of the changes in PR: https://github.com/JoshuaMulliken/wyzeapy/pull/18

It still needs to implement a token callback so that it can save the changes when the token is refreshed. It also should have a title on the 2fa input screen.