2fa (#244)
* Remove client.async_close as it is unnecessary

This is unnecessary and throws errors after https://github.com/JoshuaMulliken/wyzeapy/pull/18/commits/ae6aaa196db5a7921c5172bef9ef304144d0ecd7

* Initial 2fa support

This is preliminary 2fa support. It depends on some of the changes in PR: https://github.com/JoshuaMulliken/wyzeapy/pull/18

It still needs to implement a token callback so that it can save the changes when the token is refreshed. It also should have a title on the 2fa input screen.

* 2fa: Enable storing an updated token via callback

Given that Wyzeapy updates the token as necessary internally, we need to feed it a callback in order to store the updated tokens as it renews them so that we can actually login using the token when 2fa is enabled after a reboot/restart.

* TokenManager: carry forward the username/password

We should probably go ahead and keep the existing username/password combination when we update the entry. Previously, it was clearing these values out during the token updates.

* TokenManager: Add exception handle decorator to remove entry

This decorator should allow us to remove the wyzeapi entry when an AccessToken Error or LoginError occurs during a wyzeapy call.

* Wrap wyzeapy calls with TokenManager exception decorator.

This should wrap all the relevant calls with the decorator to catch errors with the token/login info.

* TokenManager: Fix Error handler logic

This did not work as intended, now it does.

However, this completely removes the entry for the integration so it will need to be re-installed after this happens. I'm not 100% sure that this is the most graceful way to handle this... but it also should never really happen.

* Rework re-authentication to properly notify the user and reauthenticate

This should allow for users to be notified when something goes wrong and allow them re-authenticate.

* chore: update version number, include latest beta wyzeapy

Co-authored-by: Joshua Mulliken <joshua@mulliken.net>