feat: add lock and keypad battery sensors
As we are now obtaining this data, lets provide sensors for the users so that they can easily access this data. This should also allow for the data to collect statistics through Home Assistant.
add doc strings to the class and setup function

This makes the linter happy, So go ahead and include the docstrings.
sensor: set should_poll to false - we get updates via dispatcher

There's no need for these entities to poll as the updates come via the dispatcher from another entity (lock entity in this case).
Disable the keypad_battery sensor by default

Not every user's lock will have the keypad battery available as the keypad is an additional item that is purchased separately. So, let's disable it by default and only enable it once we get an update that has the data for a keypad in it.
Fix missing space between properties


Keypad Battery Sensors: Include attribution, available and model

These basic fields should probably be included
Lock Battery Sensors: Give them a nice looking name

Common Decency would say that we should give these a nice looking name. Lets do that.
sensors: rename the list to sensors

since this won't just be for lock batteries, let's rename the list to be more generic.
lock.py: sort imports