chore: cleanup __init__.py
__init__.py: entry.data does nothing here

this is a no-effect statement, it can safely be removed.
__init__.py: correct linter complaint about user visible log messages

the linter does not want user visible log messages to end in periods.
__init__.py: linter wants lazy string formatting for log messages

appease the linter by converting these logger entries to lazy string formatting.
__init__.py: Linter does not want user visible log entries ending in .

Remove the ellipsis here to appease the linter.

__init__.py: correct linter raise-missing-from

Since we don't care about the underlying traceback here, let's just go ahead and raise from None to cleanup the traceback and appease the linter.

__init__.py: add comment to explain why 'domain' has to remain here

The linter complains that this variable is unused. If we remove it, it will remove all entities each startup for the device not being in the mac address list.