chore: cleanup old entities for unsupported devices
We're starting to need to prune old entities. This function should ensure that only supported devices are left in the device registry and that any entities that are not attached to a device in the registry are removed.
chore: cleanup outdated entities

We have been accruing some outdated entities, lets go ahead and do some cleanup.

Remove any of our entities that are no longer attached to a device.
Remove any entities who do not have a domain listed in PLATFORMS
Remove the old Notification Toggle Entity