HA-WyzeAPI 2021.11.1 (#289)
* feat: LightEntity: add IP, RSSI, SSID to state attributes (#261)Add IP, RSSI, SSID to the LightEntity's state attributes to support feature request: https://github.com/JoshuaMulliken/ha-wyzeapi/issues/145

* chore: bump version number

* fix: add self.async_schedule_update_ha_state() to switch

* chore: bump version number

* feat: add lock and keypad battery sensors

As we are now obtaining this data, lets provide sensors for the users so that they can easily access this data. This should also allow for the data to collect statistics through Home Assistant.
add doc strings to the class and setup function

This makes the linter happy, So go ahead and include the docstrings.
sensor: set should_poll to false - we get updates via dispatcher

There's no need for these entities to poll as the updates come via the dispatcher from another entity (lock entity in this case).
Disable the keypad_battery sensor by default

Not every user's lock will have the keypad battery available as the keypad is an additional item that is purchased separately. So, let's disable it by default and only enable it once we get an update that has the data for a keypad in it.
Fix missing space between properties


Keypad Battery Sensors: Include attribution, available and model

These basic fields should probably be included
Lock Battery Sensors: Give them a nice looking name

Common Decency would say that we should give these a nice looking name. Lets do that.
sensors: rename the list to sensors

since this won't just be for lock batteries, let's rename the list to be more generic.
lock.py: sort imports

* feat: add battery sensor for Wyze Cam Outdoor

If the switch is a Wyze Cam Outdoor (WVOD1) attach a battery sensor entity to it.

* chore: store the UUID in the config_entry vice ini file

The wyze_config.ini thing really annoyed me, lets go ahead and store it in the config_entry instead like the rest of our data. This commit also addresses how we are updating data in our config entry. Previously, we were hardcoding all of the values into the function call. This rewrite persists any pre-existing data without hardcoding. This means that we can add/modify data by just making calls against that specific field and the others will be left untouched.

* chore: cleanup __init__.py

__init__.py: entry.data does nothing here

this is a no-effect statement, it can safely be removed.
__init__.py: correct linter complaint about user visible log messages

the linter does not want user visible log messages to end in periods.
__init__.py: linter wants lazy string formatting for log messages

appease the linter by converting these logger entries to lazy string formatting.
__init__.py: Linter does not want user visible log entries ending in .

Remove the ellipsis here to appease the linter.

__init__.py: correct linter raise-missing-from

Since we don't care about the underlying traceback here, let's just go ahead and raise from None to cleanup the traceback and appease the linter.

__init__.py: add comment to explain why 'domain' has to remain here

The linter complains that this variable is unused. If we remove it, it will remove all entities each startup for the device not being in the mac address list.

* feat: fire event when there is a new camera event

If the last_event_ts has changed since the last update on the camera switches, lets go ahead and fire an event. HA prefixes the events with the integration name. However, the name and data structure will be:

name: wyze_camera_event

data will be provided as a dictionary:
device_name: nickname of device
device_mac : mac address of device
ai_tag_list: list of any AI tags that were provided
tag_list: list of any tags that were provided
event_screenshot: url containing the screenshot
event_video: url containing the video

* bugfix: persist existing wyze.ini uuid for a few versions

In order to avoid creating new entities, we'll need to read the value from the wyze.ini file for a few versions to make sure that everyone's config gets updated.

We also need to pass in a UUID when logging in.

* chore: stop storing the notification UUID completely

Since this doesn't really need to be stored, lets just use a static name for this single entity. This will leave a dangling entity, but that will be cleaned up in a later commit.

* bug: remove lingering code from storing UUID in config

This was causing the toggle to be auto-cleaned out of the device registry.

* #225 Add hard refresh to the doc (#268)

* feat: update to bump wyzeapy

* docs: update changelog

* chore(version): bump version in manifest

* chore(version): update version

Co-authored-by: Joe Schubert <yoinx@graffiti.net>
Co-authored-by: filename.exe <filename.exe@gmail.com>