Frontend update to support new login method of keyid / apikey (#497)
This commit includes a series of updates to the authentication system of the application. The major change is the update to accept the new keyid/apikey login method. This enhancement is significant and requires the updates submitted for the wyzeapy changes to support the login. These changes are primarily for the frontend of the application.

In addition to this, several updates have been made to different files to improve the overall functionality:

1. __init__.py: An unnecessary item was removed from the import section to streamline the code.

2. config_flow.py: This file saw multiple updates. The first update was made to improve the formatting of the code to enhance readability and maintainability. Subsequent updates were made to improve the functionality of the configuration flow.

3. const.py: This file was updated as part of the ongoing improvements.

4. token_manager.py: Changes were made to this file to manage tokens more efficiently.

5. manifest.json: The manifest.json file was updated to include the correct wyzeapy version. This is crucial to ensure that the application is using the correct version of wyzeapy.

Co-authored-by: Katie Mulliken <katie@mulliken.net>